Return type set as int for ten2eight and twelve2eight functions

+++ bms/temp/temp.ino
@@ -23,7 +23,7 @@ void canSniff(const CAN_message_t &msg) {
   } Serial.println();
 }
 
-float ten2eight(int tenBit){
+int ten2eight(int tenBit){
   int eightBit = tenBit >> 2;   
   return eightBit;
 }
@@ -55,7 +55,7 @@ void loop() {
   Can0.events();
 
   pot.val = analogRead(0);
-  pot.tx = (int)ten2eight(pot.val);
+  pot.tx = ten2eight(pot.val);
 
   static uint32_t timeout = millis();
   if ( millis() - timeout > 8 ) { // send random frame every 8ms
