+++ sensors/linear_pot/linear_pot.ino
@@ -19,7 +19,7 @@ void canSniff(const CAN_message_t &msg) {
   } Serial.println();
 }
 
-float ten2eight(int tenBit) {
+int ten2eight(int tenBit) {
   int eightBit = tenBit >> 2;
   return eightBit;
 }
@@ -45,7 +45,7 @@ void loop() {
 
   // Convert potentiometer reads from Pin0 from 10bit to 8bit to send via CAN
   pot.val = analogRead(A0);
-  pot.msg = (int)ten2eight(pot.val);
+  pot.msg = ten2eight(pot.val);
   Serial.print(pot.msg);
   delay(250);
